backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
audioread==3.0.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
decorator==5.2.1
exceptiongroup==1.3.0
facenet-pytorch==2.6.0
fastapi==0.118.2
filelock==3.20.0
fsspec==2025.9.0
h11==0.16.0
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
MarkupSafe==3.0.3
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
numba==0.62.1
numpy==1.26.4
opencv-python-headless==4.8.0.76
packaging==25.0
pillow==10.2.0
platformdirs==4.5.0
pooch==1.8.2
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.20
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
huggingface-hub